+++ shaders/lib/sharpening.glsl
@@ -81,9 +81,8 @@ void doSharpening(inout vec3 color) {
 		
 	#endif
 	
-	float cameraVel = length(cameraPosition - previousCameraPosition);
-	cameraVel = min(cameraVel, 0.1);
-	color = mix(color, blur, (SHARPEN_AMOUNT / 5.0 + cameraVel * SHARPEN_VEL_ADDITION * 2.0) * -1.0);
+	float sharpenAmount = SHARPEN_AMOUNT * 0.1 + sharpenVelocityFactor * SHARPEN_VEL_ADDITION;
+	color = mix(color, blur, sharpenAmount * -1.0); // exaggerate the difference between the image and the blurred image
 	//color = blur;
 	
 }